extract validation to separate file
